requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aiohttp==3.9.0
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
atlastk==0.13.3
attrs==23.1.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dictdiffer==0.9.0
frozenlist==1.4.0
html5tagger==1.3.0
httptools==0.6.1
idna==3.4
iso8601==1.1.0
multidict==6.0.4
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
PyJWT==2.8.0
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sanic==23.6.0
sanic-ext==23.6.0
sanic-routing==23.6.0
SQLAlchemy==2.0.23
tomlkit==0.12.3
tortoise-orm==0.20.0
tracerite==1.1.1
typing_extensions==4.8.0
ujson==5.8.0
uvloop==0.19.0
websockets==12.0
yarl==1.9.2
